fix: renamed toggle grooup to toggle button group
