+++ packages/design-system/src/ToggleButtonGroup/index.ts
@@ -1,0 +1,2 @@
+export * from "./ToggleButtonGroup";
+export * from "./ToggleButtonGroup.types";
