--- packages/design-system/src/ToggleGroup/index.ts
@@ -1,2 +1,0 @@
-export * from "./ToggleGroup";
-export * from "./ToggleGroup.types";
